# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Testing api",
        "version": "1.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "User log In and return generated token",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLoginRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "User logout",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/registration": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registration",
                "description": "User registration",
                "operationId": "registration",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRegistrationRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRegistrationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/call-event": {
            "post": {
                "tags": [
                    "CallEvent"
                ],
                "summary": "Call event",
                "description": "Call event",
                "operationId": "call-event",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CallEventRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CallEventRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "UserLoginRequest": {
                "title": "User Log In request",
                "description": "User Login request body data",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "Email",
                        "type": "string",
                        "example": "test@bk.ru"
                    },
                    "password": {
                        "title": "password",
                        "description": "User password",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "UserRegistrationRequest": {
                "title": "User Registration request",
                "description": "User Registration request body data",
                "required": [
                    "email",
                    "name",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name",
                        "type": "string",
                        "example": "username"
                    },
                    "email": {
                        "title": "email",
                        "description": "Email",
                        "type": "string",
                        "example": "test@bk.ru"
                    },
                    "password": {
                        "title": "password",
                        "description": "User password",
                        "type": "string",
                        "example": "password"
                    },
                    "password_confirmation": {
                        "title": "password_confirmation",
                        "description": "User password_confirmation",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "CallEventRequest": {
                "title": "Call Event request",
                "description": "Call Event request body data",
                "required": [
                    "call_id",
                    "from",
                    "to",
                    "event_type",
                    "timestamp"
                ],
                "properties": {
                    "call_id": {
                        "title": "call_id",
                        "description": "Call id",
                        "type": "string",
                        "example": "445436456"
                    },
                    "from": {
                        "title": "from",
                        "description": "from number",
                        "type": "string",
                        "example": "994550000000"
                    },
                    "to": {
                        "title": "to",
                        "description": "to number",
                        "type": "string",
                        "example": "994550000000"
                    },
                    "event_type": {
                        "title": "event_type",
                        "description": "event_type",
                        "type": "string",
                        "example": "call_ended"
                    },
                    "timestamp": {
                        "title": "timestamp",
                        "description": "timestamp",
                        "type": "string",
                        "example": "2025-10-11 11:12"
                    },
                    "duration": {
                        "title": "duration",
                        "description": "duration",
                        "type": "integer",
                        "example": "10",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "CallEvent",
            "description": "CallEvent"
        }
    ]
}